Added chat logging to blackfire
100% compatible with blackfire1.6, you don't even have to replace your application support directory

+++ BlackFire/BFChatWindowController.m
@@ -322,7 +322,7 @@
 
 		if( remoteFriend.gameID > 0 )
 		{
-			displayImage = [[BFGamesManager sharedGamesManager] imageForGame:remoteFriend.gameID];
+			displayImage = [[BFGamesManager sharedGamesManager] imageForGame:(unsigned int)remoteFriend.gameID];
 			if( [statusString length] > 0 )
 			{
 				statusString = [NSString stringWithFormat:@"%@, ",statusString];
